+++ src/main/java/PodoeMarket/podoemarket/admin/dto/request/PaymentStatusRequestDTO.java
@@ -1,0 +1,14 @@
+package PodoeMarket.podoemarket.admin.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PaymentStatusRequestDTO {
+    private Boolean checked;
+}
