+++ src/main/java/PodoeMarket/podoemarket/common/config/EmailConfig.java
@@ -1,4 +1,4 @@
-package PodoeMarket.podoemarket.config;
+package PodoeMarket.podoemarket.common.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
