+++ src/main/java/PodoeMarket/podoemarket/common/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package PodoeMarket.podoemarket.config;
+package PodoeMarket.podoemarket.common.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
