+++ src/main/java/PodoeMarket/podoemarket/common/config/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package PodoeMarket.podoemarket.config.jwt;
+package PodoeMarket.podoemarket.common.config.jwt;
 
 import lombok.Getter;
 import lombok.Setter;
