+++ src/main/java/PodoeMarket/podoemarket/common/entity/OrderItemEntity.java
@@ -1,10 +1,11 @@
-package PodoeMarket.podoemarket.entity;
+package PodoeMarket.podoemarket.common.entity;
 
-import PodoeMarket.podoemarket.entity.type.PlayType;
-import PodoeMarket.podoemarket.entity.type.ProductStatus;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 
 import java.time.LocalDateTime;
@@ -14,58 +15,37 @@ import java.util.List;
 import java.util.UUID;
 
 @Entity
-@Table(name = "product")
+@Table(name = "orderItem")
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProductEntity {
+public class OrderItemEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID id;
 
     @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String writer;
-
-    @Column(nullable = false)
-    private String filePath;
-
-    @Column
-    private String imagePath;
-
-    @Column(nullable = false)
     @ColumnDefault("false")
-    private boolean script; // 대본권 판매 여부
+    private boolean script; // 대본권 구매 여부
 
     @Column(nullable = false)
     @ColumnDefault("0")
     private int scriptPrice;
 
     @Column(nullable = false)
-    @ColumnDefault("false")
-    private boolean performance; // 공연권 판매 여부
+    @ColumnDefault("0")
+    private int performanceAmount;
 
     @Column
     @ColumnDefault("0")
     private int performancePrice;
 
-    @Column
-    private String descriptionPath;
-
-    @Enumerated(EnumType.STRING)
-    @Column
-    private PlayType playType;
-
-    @Column
-    private String plot;
-
-    // 관리자(심사 주체) 확인 여부
-    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private ProductStatus checked = ProductStatus.WAIT;
+    private int totalPrice;
+
+    @Column(nullable = false)
+    private String title;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
@@ -79,17 +59,35 @@ public class ProductEntity {
         createdAt = now;
         updatedAt = now;
     }
+
     @PreUpdate // db에 entity가 업데이트되기 직전에 실행
     protected void onUpdate() { updatedAt = LocalDateTime.now(ZoneId.of("Asia/Seoul")); }
 
-    // user : product = 1 : N
+    // orders : orderItem = 1 : N
+    @ManyToOne(targetEntity = OrdersEntity.class)
+    @JoinColumn(name = "order_id", nullable = false)
+    @JsonIgnore
+    private OrdersEntity order;
+
+    // product : orderItem = 1 : N
+    @ManyToOne(targetEntity = ProductEntity.class)
+    @JoinColumn(name = "product_id")
+    @JsonIgnore
+    private ProductEntity product;
+
+    // user : orderItem = 1 : N
     @ManyToOne(targetEntity = UserEntity.class)
-    @JoinColumn(name = "user_id")
+    @JoinColumn(name = "user_id", nullable = false)
     @JsonIgnore
     private UserEntity user;
 
-    // product : orderItem = 1 : N
-    @OneToMany(mappedBy = "product")
+    // orderItem : performanceDate = 1 : N
+    @OneToMany(mappedBy = "orderItem")
     @JsonIgnore
-    private List<OrderItemEntity> orderItem = new ArrayList<>();
+    private List<PerformanceDateEntity> performanceDate = new ArrayList<>();
+
+    // orderItem : applicant = 1 : 1
+    @OneToOne(mappedBy = "orderItem")
+    @JsonIgnore
+    private ApplicantEntity applicant;
 }
