+++ src/main/java/PodoeMarket/podoemarket/common/entity/PerformanceDateEntity.java
@@ -1,4 +1,4 @@
-package PodoeMarket.podoemarket.entity;
+package PodoeMarket.podoemarket.common.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,30 +6,25 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Entity
-@Table(name = "applicant")
+@Table(name = "performanceDate")
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ApplicantEntity {
+public class PerformanceDateEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID id;
 
     @Column
-    private String name;
+    private LocalDateTime date;
 
-    @Column
-    private String phoneNumber;
-
-    @Column
-    private String address;
-
-    // orderItem : applicant = 1 : 1
-    @OneToOne(targetEntity = OrderItemEntity.class)
-    @JoinColumn(name = "orderItem_id")
+    // orderItem : performanceDate = 1 : N
+    @ManyToOne(targetEntity = OrderItemEntity.class)
+    @JoinColumn(name  = "orderItem_id", nullable = false)
     private OrderItemEntity orderItem;
 }
