+++ src/main/java/PodoeMarket/podoemarket/common/entity/ProductEntity.java
@@ -1,11 +1,10 @@
-package PodoeMarket.podoemarket.entity;
+package PodoeMarket.podoemarket.common.entity;
 
+import PodoeMarket.podoemarket.common.entity.type.PlayType;
+import PodoeMarket.podoemarket.common.entity.type.ProductStatus;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
 
 import java.time.LocalDateTime;
@@ -15,37 +14,58 @@ import java.util.List;
 import java.util.UUID;
 
 @Entity
-@Table(name = "orderItem")
+@Table(name = "product")
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class OrderItemEntity {
+public class ProductEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID id;
 
     @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String writer;
+
+    @Column(nullable = false)
+    private String filePath;
+
+    @Column
+    private String imagePath;
+
+    @Column(nullable = false)
     @ColumnDefault("false")
-    private boolean script; // 대본권 구매 여부
+    private boolean script; // 대본권 판매 여부
 
     @Column(nullable = false)
     @ColumnDefault("0")
     private int scriptPrice;
 
     @Column(nullable = false)
-    @ColumnDefault("0")
-    private int performanceAmount;
+    @ColumnDefault("false")
+    private boolean performance; // 공연권 판매 여부
 
     @Column
     @ColumnDefault("0")
     private int performancePrice;
 
-    @Column(nullable = false)
-    private int totalPrice;
+    @Column
+    private String descriptionPath;
 
+    @Enumerated(EnumType.STRING)
+    @Column
+    private PlayType playType;
+
+    @Column
+    private String plot;
+
+    // 관리자(심사 주체) 확인 여부
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private String title;
+    private ProductStatus checked = ProductStatus.WAIT;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
@@ -59,35 +79,17 @@ public class OrderItemEntity {
         createdAt = now;
         updatedAt = now;
     }
-
     @PreUpdate // db에 entity가 업데이트되기 직전에 실행
     protected void onUpdate() { updatedAt = LocalDateTime.now(ZoneId.of("Asia/Seoul")); }
 
-    // orders : orderItem = 1 : N
-    @ManyToOne(targetEntity = OrdersEntity.class)
-    @JoinColumn(name = "order_id", nullable = false)
-    @JsonIgnore
-    private OrdersEntity order;
-
-    // product : orderItem = 1 : N
-    @ManyToOne(targetEntity = ProductEntity.class)
-    @JoinColumn(name = "product_id")
-    @JsonIgnore
-    private ProductEntity product;
-
-    // user : orderItem = 1 : N
+    // user : product = 1 : N
     @ManyToOne(targetEntity = UserEntity.class)
-    @JoinColumn(name = "user_id", nullable = false)
+    @JoinColumn(name = "user_id")
     @JsonIgnore
     private UserEntity user;
 
-    // orderItem : performanceDate = 1 : N
-    @OneToMany(mappedBy = "orderItem")
+    // product : orderItem = 1 : N
+    @OneToMany(mappedBy = "product")
     @JsonIgnore
-    private List<PerformanceDateEntity> performanceDate = new ArrayList<>();
-
-    // orderItem : applicant = 1 : 1
-    @OneToOne(mappedBy = "orderItem")
-    @JsonIgnore
-    private ApplicantEntity applicant;
+    private List<OrderItemEntity> orderItem = new ArrayList<>();
 }
