+++ src/main/java/PodoeMarket/podoemarket/common/entity/type/PlayType.java
@@ -1,0 +1,6 @@
+package PodoeMarket.podoemarket.common.entity.type;
+
+public enum PlayType {
+    LONG,
+    SHORT
+}
