+++ src/main/java/PodoeMarket/podoemarket/common/repository/PerformanceDateRepository.java
@@ -1,6 +1,6 @@
-package PodoeMarket.podoemarket.repository;
+package PodoeMarket.podoemarket.common.repository;
 
-import PodoeMarket.podoemarket.entity.PerformanceDateEntity;
+import PodoeMarket.podoemarket.common.entity.PerformanceDateEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
