+++ src/main/java/PodoeMarket/podoemarket/common/repository/RefundRepository.java
@@ -1,6 +1,6 @@
-package PodoeMarket.podoemarket.repository;
+package PodoeMarket.podoemarket.common.repository;
 
-import PodoeMarket.podoemarket.entity.RefundEntity;
+import PodoeMarket.podoemarket.common.entity.RefundEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RefundRepository extends JpaRepository<RefundEntity, Long> {
