+++ src/main/java/PodoeMarket/podoemarket/controller/ProductController.java
@@ -1,11 +1,12 @@
 package PodoeMarket.podoemarket.controller;
 
+import PodoeMarket.podoemarket.common.entity.ProductEntity;
+import PodoeMarket.podoemarket.common.entity.UserEntity;
 import PodoeMarket.podoemarket.dto.*;
 import PodoeMarket.podoemarket.dto.response.ProductListDTO;
 import PodoeMarket.podoemarket.dto.response.ResponseDTO;
 import PodoeMarket.podoemarket.dto.response.ScriptListDTO;
-import PodoeMarket.podoemarket.entity.*;
-import PodoeMarket.podoemarket.entity.type.PlayType;
+import PodoeMarket.podoemarket.common.entity.type.PlayType;
 import PodoeMarket.podoemarket.service.ProductService;
 import PodoeMarket.podoemarket.service.S3Service;
 import lombok.RequiredArgsConstructor;
