+++ src/main/java/PodoeMarket/podoemarket/dto/response/OrderScriptDTO.java
@@ -1,7 +1,7 @@
 package PodoeMarket.podoemarket.dto.response;
 
-import PodoeMarket.podoemarket.entity.type.PlayType;
-import PodoeMarket.podoemarket.entity.type.ProductStatus;
+import PodoeMarket.podoemarket.common.entity.type.PlayType;
+import PodoeMarket.podoemarket.common.entity.type.ProductStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
