--- src/main/java/PodoeMarket/podoemarket/entity/type/PlayType.java
@@ -1,6 +1,0 @@
-package PodoeMarket.podoemarket.entity.type;
-
-public enum PlayType {
-    LONG,
-    SHORT
-}
