+++ src/main/java/PodoeMarket/podoemarket/service/RegisterService.java
@@ -1,7 +1,7 @@
 package PodoeMarket.podoemarket.service;
 
-import PodoeMarket.podoemarket.entity.ProductEntity;
-import PodoeMarket.podoemarket.repository.ProductRepository;
+import PodoeMarket.podoemarket.common.entity.ProductEntity;
+import PodoeMarket.podoemarket.common.repository.ProductRepository;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import lombok.RequiredArgsConstructor;
