+++ src/main/java/PodoeMarket/podoemarket/service/UserService.java
@@ -1,7 +1,7 @@
 package PodoeMarket.podoemarket.service;
 
-import PodoeMarket.podoemarket.entity.UserEntity;
-import PodoeMarket.podoemarket.repository.UserRepository;
+import PodoeMarket.podoemarket.common.entity.UserEntity;
+import PodoeMarket.podoemarket.common.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
